requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
confusable-homoglyphs==3.2.0
Django==3.1.6
django-bootstrap4==2.3.1
django-crispy-forms==1.11.0
django-registration==3.1.1
django-simple-history==2.12.0
importlib-metadata==2.1.1
pkg-resources==0.0.0
psycopg2==2.8.6
pytz==2021.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
zipp==3.4.0
